pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.22.0",
    "environs>=14.3.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["My_project"]

[project.scripts]
bot = "My_project.main:"
